+++ src/data/layout/sort/mod.rs
@@ -3,23 +3,16 @@
 //! Sorting functionality.
 //
 
-// no items defined
+// implementations
 mod generic;
 mod primitives;
+#[cfg(feature = "alloc")]
+mod impls_alloc;
 
 mod definition; // Sort
-
-#[cfg(feature = "alloc")]
-crate::items! {
-    mod definition_alloc; // SortAlloc
-    mod impls_alloc;
-}
 
 crate::structural_mods! { // _mods
     _mods {
         pub use super::definition::*;
-
-        #[cfg(feature = "alloc")]
-        pub use super::definition_alloc::*;
     }
 }